Test Scene
Added place holder assets for the skill sprites.

Fixed a bug that only displayed the skill icon for the most recent skill used when removing a status effect from the player.

Fixed a bug that didn't show the correct skill icon used for status effect when the player clicks on a skill.