Test Scene
The enemy health bars now disable their images and texts instead of deactivating the whole game object. This is so that certain status effects afflicted on an enemy won't encounter weird behaviors.

Changed the positions of the player damage and status effect texts.

Fixed a bug where an enemy would also receive the player's status effects.

Added a fade out to the status icon image in the player's status effect text animation.