Test Scene
Adjusted the vertical layout group for the text animations so that they don't jitter as much when multiple text objects are parented into the layout object.

The player's health and mana bars now lerp back to their max values after the player respawns instead of just instantly resetting.

added the "Taking Damage" boolean value inside of the "ModifyHealth" and "IncreaseHealth" functions instead of setting the value outside of the class.

The screen now fades out when the player dies and then fades back in when the player respawns.