Test Scene
Removed the Toggle Health bar function for the enemy from update and instead call the function when necessary instead.

The animation of the invalid text will now be replayed if the player continues to put in input for a target that's too far away, a non existing target, or with an insufficient amount of mana.

Made two new folders in the prefabs folder to hold specific prefab types for organization purposes.