Test Scene
Added two new status effects: Haste, and Blood And Sinew. The definition of these states still needs to be written.

Adjusted the UI for the main menu scene to scale with the game window and also changed the button texts to text mesh pro.

Changed the remainding default UI texts to utilize text mesh pro.

Added an object pooler script. So far only works with player damage for testing.

Added a new skill: Blood And Sinew.

Added a separate effect state on the skill script for enemy states.