Test Scene
Added custom cursors for default navigation and when hovering over an enemy.

Fixed a bug where the enemy target icon would still be active when moving outside of the target range.

Fixed a bug where the fade in and fade out screen would not reactivate after dying and respawing after the first time.