Test Scene
Added a new sprite for the eemy target image.

Added a bare bones sound manager only including the singleton pattern.

Adjusted the experience bar such that when the player receives a level, the fill amount will start from the beginning.

Adjusted the text size for the status effect and the description of the status.