Test Scene
Added a particle effect for when a status effect is removed from a player or an enemy.

Added more particle effects into the game.

Changed the low health animation from legacy to default.

Fixed an issue where the low health animation would only stop halfway for the player if they were above a certain health threshold.

Adjusted the attack range for the player and the mushroom enemy.