Test Scene
Adjusted enemies health bar size and the size of the status icons that are placed under the enemies healh bars.

Added a hit particle to the object pooler.

Changed the animation event of the player's attack animation so that the damage is dealt the moment the player's weapon connects.

Changed the MushroomAnimations script to EnemyAnimations for future enemies.