Test Scene
Added functionality to the Storm Thrust skill.

Added a new animation for the player when using Storm Thrust.

Added a separate particle effect when the doom status effect is removed from the player.

Added a hit particle object into the object pooler.

Added hit particles for the swift strike ability.