Test Scene
Fixed a bug where the player was able to activate the same skill if another one was in the skill bar. Duplicate skills will now share the same cooldown.

Fixed a bug where the player was not able to remove a skill from the skill bar for certain skills.

Added some functionality to the skill menu. A skill can now be placed from the skill menu onto the skill bar and also be removed.

Removed some obsolete variables from the health script.

Changed the script name for the LowEnemyHPAnimation to LowHpAnimation.