Test Scene
When a player levels up and reaches the level requirement for a skill, the new skill will be added to the skillbar.

Added an image to all skills to represent cooldowns.

Skills that are not yet learned cannot be removed from the skills menu.

Skills that are on a cool down will now show in the skill menu when the menu is both opened and closed.