Test Scene
Passive skills can now be learned.

Added the following passive skill bonuses: HP, MP, Strength, Defense, Intelligence, and Critical.

Changed the text color of the status effect bonus and duration on the skills that provide status effects to the same color as the status icon display name.

Adjusted the font size of the texts in the character menu as well as changing the color of the hp and mp texts to match their UI's text color.

Fixed a bug where the skill menu would reactivate if the player tries to change to a different menu.

Added an underscore to the skill names in the skill text boxes.