Test Scene
The character menu now updates stat values even while it is opened.

Stat values in the character menu change color based on bonuses such as increased strength from a skill.

Added a duration variable to the particle effects so that they play out the entire particle animation before returning to the object pool.

Added a separate transform for the status effect text for the player so that it does not interfere with the alignment of the player damage/heal texts when they are displayed.

Added an animation to the Whirlwind Slash skill.

Fixed a bug where the mana text and value would not clamp when fully restored.

Changed the text color of the added effects affix on the character's skills.

Added an underscore to the skills in the skills panel.