Test Scene
Centered the enemy damage and status effect texts.

Added the Z and X hot keys to the item panel.

Added an item manager that allows the player to press Z or X to use an item corresponding to the keys.

The character's rotation for whirlwind slash will play a few frames after the animation for the skill begins.

The character will return to the original rotation they were at before using whirlwind slash.

Fixed a bug where the character would not play their death animation when they died while currently using a skill or casting a spell.