Test Scene
Added and implemented a new enemy: Bee

Added a skill animation for when the player uses a self inducing status effect.

Added a particle effect for the potion and ether items, and a particle effect for the tenacity skill.

Fixed a bug where the player's animations would break if they tried to use the tenacity skill along with any other skill at once.

Adjusted the skill text in the description panels by adding an extra break from the mp cost and the power from the cooldown and the cast time as well as change mana to mp cost and changed the order of the power, mp cost, cast time, and cooldown texts.