Test Scene
Changed all of the enemy materials from alpha to opaque and added a script that changes their materials to an alpha material so they can fade away when they die.

Readjusted the damage radius for the rectangle shape so that the collider matches the size of the recttransform.

Added a new skill for the Bee enemy: Stunning Stinger

Added a casting animation for the Bee.

Items now become unusable when the player is stunned.

Fixed a bug where an enemy wouldn't play their death animation and disappear if they didn't have any skills attached.

Adjusted the move animation for the Bunnykins enemy so that it doesn't jump higher than its health bar.