Level 1
Finished laying out the stage.

Added a new skill for the player: Evil's End.

Changed the player's class from warrior to knight.

Added a new status icon: Poison

Added enemies to the stage.

Modified the enemy AI script to chose to apply a normal attack if they don't have any skills set up.

Fixed a few typos for the player's animator fields.

Reduced the texture size of all of the textures that are currently in use in the scene.