Level 1 updates and bug fixes
Fixed a bug where the cursor would change to the attack cursor when the player was close enough to the enemy but would not be able to click on the target.

Updated the tenacity particle effect.

Made a new particle effect for the enemy cast.

Swapped the status effect text and the damage text holder prefabs to the correct transforms.

Updated the poison status effect icon.

Adjusted the status effect text display when applied to the player and enemies.