Level 1
Adjusted lighting in the scene and added a new terrain to work with.

Fixed an issue with the enemy animations ignoring animation events.

Fixed a bug where the skill learned text would throw off warnings about no animator controllers being attached.

Added an actual casting animation to the bee and made the clip loop.

Adjusted the skill learned text to appear more legible when shown on the screen.