Level 1
The player's rotation for whirlwind slash now stops at the exact rotation the player was in before the skill's activation.

The damage radius for the enemy's skills will now hit the player if they stood within the radius after the enemy finished casting their skill.

Adjusted the skill learned text to appear more legible.

Added more environmental assets to the game.

Fixed a bug where the player would still be in the walking animation when they were affected by the stun status effect.