Bug Fixing & Balancing
Fixed a bug where the Shadow Priest's Diabolic Lightning skill wouldn't affect the second enemy in a group if the first enemy dies from it.

Fixed a bug where enemies in the enemy trigger zone wouldn't die properly or give experience/coins/skill points if they were killed without the player having stepped into the trigger.

Changed the stats on all bosses, the power of the Earth Spirit's Slag skill, and added increased defense to all of the Shadow Priest's off-hand equipment.

Decreased the cast speed of Shatter and decreased the cast speed and the mana cost of Soul Pierce.