Bug fixes & Balancing
Fixed a bug where the Slowed status effect would result in strange behavior when inflicted on a boss.

Reduced the amount of damage bosses take from damage dealing status effect and also made them immune to the Stun status effect.

Reduced the percentage of the equipment abilities that give status effects on enemies from 10% to 5%.